Rename RepoSearch component to general Search
